+++ mobile/android/android-components/components/feature/top-sites/src/main/java/mozilla/components/feature/top/sites/DefaultTopSitesStorage.kt
@@ -102,6 +102,7 @@ class DefaultTopSitesStorage(
             try {
                 providerTopSites = topSitesProvider
                     .getTopSites(allowCache = true)
+                    .filter { providerConfig.providerFilter?.invoke(it) ?: true }
                     .take(numSitesRequired)
                     .take(providerConfig.maxThreshold - pinnedSites.size)
                 topSites.addAll(providerTopSites)